--- config/index.js
@@ -1,6 +1,0 @@
-import { name } from '../package.json';
-import i18n     from './i18n'
-export default {
-  i18n,
-  name
-}
