--- form/outputs.js
@@ -1,10 +1,0 @@
-import outputvectorlayer from "../components/OutputVectorLayer.vue";
-import outputrasterlayer from "../components/OutputRasterLayer.vue";
-import outputfile        from "../components/OutputFile.vue";
-
-export default {
-  outputvectorlayer,
-  outputrasterlayer,
-  outputfile,
-  outputhtml: outputfile,
-}
